requirements.txt
mcp[cli]
fastmcp
tavily-python
httpx
beautifulsoup4
python-dotenv
